added logout and pathfiles
